requirements.txt
Django==1.7.4
Pillow==2.7.0
Unidecode==0.04.17
dj-database-url==0.3.0
dj-static==0.0.6
django-endless-pagination==2.0
django-forms-bootstrap==3.0.1
django-toolbelt==0.0.1
django-uuslug==1.0.3
gunicorn==19.2.1
psycopg2==2.6
python-slugify==0.1.0
pytz==2014.10
static3==0.5.1
wsgiref==0.1.2
